+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - updated most of of the dependencies
+- :boom: changed to own reactstrap fork with react 19 support
 
 ## [5.3.1] - 2026-02-04